move section picker + token to section module
