+++ src/packages/core/section/index.ts
@@ -5,3 +5,5 @@ export * from './section-sidebar-menu/index.js';
 export * from './section-sidebar-menu-with-entity-actions/index.js';
 export * from './section-default.element.js';
 export * from './section.context.js';
+export * from './input-section/index.js';
+export * from './section-picker-modal.token.js';
