+++ src/packages/core/section/input-section/input-section.element.ts
@@ -1,7 +1,7 @@
-import { UmbInputListBaseElement } from '../input-list-base/input-list-base.js';
+import { UmbInputListBaseElement } from '../../components/input-list-base/input-list-base.js';
+import { UMB_SECTION_PICKER_MODAL } from '../section-picker-modal.token.js';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
 import { css, html, nothing, customElement, state } from '@umbraco-cms/backoffice/external/lit';
-import { UMB_SECTION_PICKER_MODAL } from '@umbraco-cms/backoffice/modal';
 import type { ManifestSection } from '@umbraco-cms/backoffice/extension-registry';
 import { umbExtensionsRegistry } from '@umbraco-cms/backoffice/extension-registry';
